% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostname.R
\name{change_hostname}
\alias{change_hostname}
\title{Change the hostname}
\usage{
change_hostname(hostname)
}
\arguments{
\item{hostname}{A new hostname, a string starting with "https://"}
}
\value{
The new hostname
}
\description{
Change the hostname
}
