% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_json.R
\name{differ_length}
\alias{differ_length}
\title{Recognizes the length of input to apply or lapply a function}
\usage{
differ_length(input, func)
}
\arguments{
\item{input}{A single element or a list}

\item{func}{A function to be applied to the input}
}
\value{
The result of the function applied in the differentiated way
}
\description{
Recognizes the length of input to apply or lapply a function
}
