% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_string}
\alias{format_string}
\title{Format a string}
\usage{
format_string(text)
}
\arguments{
\item{text}{A character vector}
}
\value{
A character vector formatted in the required way
}
\description{
Format a string
}
