% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_uid}
\alias{generate_uid}
\title{Generate a counting function
used for assigning unique identifiers}
\usage{
generate_uid()
}
\value{
The counting function
}
\description{
Generate a counting function
used for assigning unique identifiers
}
