% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference_classes.R
\name{load_reference_classes}
\alias{load_reference_classes}
\title{Load reference classes for a template and its components}
\usage{
load_reference_classes(template_id)
}
\arguments{
\item{template_id}{The ID of an ORKG template}
}
\value{
reference classes for the template and its components
}
\description{
Load reference classes for a template and its components
}
\examples{
tp <- load_reference_classes("R937648")

}
