% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_orkg.R
\name{request_orkg}
\alias{request_orkg}
\title{Request the ORKG REST API}
\usage{
request_orkg(route)
}
\arguments{
\item{route}{The part of path for requesting the ORKG REST API
which follows the hostname}
}
\value{
Requested information about an ORKG object
}
\description{
Request the ORKG REST API
}
