% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_fields}
\alias{show_fields}
\title{Title}
\usage{
show_fields(ref_class)
}
\arguments{
\item{ref_class}{A reference class based on a template}
}
\value{
The string enlisting all fields for writing an instance
}
\description{
Title
}
\examples{
tp <- load_reference_classes("R937648")
show_fields(tp$measurement_scale)

}
