% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_orkg_jsonld.R
\name{to_orkg_jsonld}
\alias{to_orkg_jsonld}
\title{Turn an instance of a reference class into ORKG-harvestable JSON-LD}
\usage{
to_orkg_jsonld(instance)
}
\arguments{
\item{instance}{An instance of a reference class}
}
\value{
JSON string in JSON-LD format
}
\description{
Turn an instance of a reference class into ORKG-harvestable JSON-LD
}
\examples{
tp <- load_reference_classes("R937648")
my_instance <- tp$measurement_scale(label = "my_scale")
my_json <- to_orkg_jsonld(my_instance)

}
