% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_json.R
\name{turn_json}
\alias{turn_json}
\title{Turn an instance of a reference class into ORKG-harvestable JSON-LD}
\usage{
turn_json(instance)
}
\arguments{
\item{instance}{An instance of a reference class}
}
\value{
JSON string in JSON-LD format
}
\description{
Turn an instance of a reference class into ORKG-harvestable JSON-LD
}
\examples{
tp <- load_reference_classes("R937648")
my_instance <- tp$measurement_scale(label = "my_scale")
my_json <- turn_json(my_instance)

}
