% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_json_orkg.R
\name{turn_json_orkg}
\alias{turn_json_orkg}
\title{Turn an instance of a reference class into ORKG-harvestable JSON-LD}
\usage{
turn_json_orkg(instance)
}
\arguments{
\item{instance}{An instance of a reference class
that is an R object containing active binding}
}
\value{
JSON string
}
\description{
Turn an instance of a reference class into ORKG-harvestable JSON-LD
}
