% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_json.R
\name{df_structure}
\alias{df_structure}
\title{Writes dataframes and tuples into JSON-LD
for the internal use}
\usage{
df_structure(df, label)
}
\arguments{
\item{df}{A dataframe or a tuple}

\item{label}{A character string}
}
\value{
JSON-LD of the dataframe of a tuple
}
\description{
Writes dataframes and tuples into JSON-LD
for the internal use
}
