% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_json_orkg.R
\name{differ_type}
\alias{differ_type}
\title{Reprocess the input to be used in JSON-LD depending on its type}
\usage{
differ_type(input)
}
\arguments{
\item{input}{A dataframe, a tuple, or another basic data type}
}
\value{
The resulting R object to be used in JSON-LD
}
\description{
Reprocess the input to be used in JSON-LD depending on its type
}
