% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference_classes.R
\name{extractor_orkg}
\alias{extractor_orkg}
\title{Extractor function for the ORKG templates}
\usage{
extractor_orkg(template_id)
}
\arguments{
\item{template_id}{}
}
\description{
Extractor function for the ORKG templates
}
