% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference_classes.R
\name{load_reference_classes}
\alias{load_reference_classes}
\title{Load reference class(es) for a template and its components}
\usage{
load_reference_classes(template_id)
}
\arguments{
\item{template_id}{}
}
\value{
reference class(es)
}
\description{
Load reference class(es) for a template and its components
}
