% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\name{set_class}
\alias{set_class}
\title{Set a reference class(es) for a template and its components}
\usage{
set_class(JSON_string)
}
\arguments{
\item{JSON_string}{}
}
\value{
reference class(es)
}
\description{
Set a reference class(es) for a template and its components
}
