% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_fields}
\alias{show_fields}
\title{Show the fields of a reference class for a template
which can be used to write an instance}
\usage{
show_fields(ref_class)
}
\arguments{
\item{ref_class}{A reference class based on a template}
}
\value{
The string listing all fields for writing an instance
}
\description{
Show the fields of a reference class for a template
which can be used to write an instance
}
\examples{
tp <- load_reference_classes("R937648")
show_fields(tp$measurement_scale)

}
