% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostname.R
\name{with_host}
\alias{with_host}
\title{Title}
\usage{
with_host(...)
}
\arguments{
\item{...}{\itemize{
\item one or more arguments to be concatenated
representing the route of the URL
}}
}
\description{
Title
}
